# ENH: sparse: speedup LIL matrix slicing (was #3338)