This is a continuation of @pv's LIL enhancements PR (#3338). I rebased on master and fixed up the benchmarks for sparse slicing to include the LIL format:
```
BENCHMARK                                    BEFORE      AFTER     FACTOR
...se.NullSlice.time_3_rows(0.01, 'lil')   229.53μs    60.48μs   0.26350105x
...se.NullSlice.time_3_rows(0.05, 'lil')   285.01μs    61.80μs   0.21684491x
...ullSlice.time_10000_rows(0.05, 'lil')   554.35ms    18.44ms   0.03326798x
...ullSlice.time_10000_rows(0.01, 'lil')   431.39ms     7.70ms   0.01785378x
```