@larsmans 2016-02-01
@rgommers and @jnothman already gave a :+1: at #3338 and I presume @pv is also still happy with this code; the benchmark code should be uncontroversial. Merging as the diff looks fine.