@pv 2016-02-01
I have no memory of this code, but looks OK.